server/requirements.txt
websocket-server==0.6.4
websocket-client==1.8.0